refactor: erraze temporary data and pages
