+++ src/Components/header/header.tsx
@@ -2,7 +2,6 @@ import './header.scss';
 
 import type { FC } from 'react';
 import React from 'react';
-import { Link } from 'react-router-dom';
 
 import pngUrl from '../../assets/png/logo.png';
 
@@ -10,28 +9,6 @@ const Header: FC = () => {
   return (
     <div className="header-menu">
       <img src={pngUrl} alt="logo"></img>
-      <ul>
-        <li>
-          <Link to="/" className="header-link">
-            Main
-          </Link>
-        </li>
-        <li>
-          <Link to="/lobby" className="header-link">
-            Lobby
-          </Link>
-        </li>
-        <li>
-          <Link to="/game" className="header-link">
-            Game
-          </Link>
-        </li>
-        <li>
-          <Link to="/game-result" className="header-link">
-            Game-result
-          </Link>
-        </li>
-      </ul>
       <div className="header-menu_bottom"></div>
     </div>
   );
