--- src/Pages/Test-page/Test-page.scss
@@ -1,19 +1,0 @@
-.test-page {
-  position: relative;
-  max-width: 1440px;
-  margin: 0 auto;
-  display: flex;
-  flex-direction: column;
-  min-height: calc(100vh - 134px);
-}
-
-.test-page-wrapper {
-  width: 100%;
-  height: 100%;
-  margin: 50px auto;
-  display: flex;
-  flex-direction: row;
-  column-gap: 10px;
-  row-gap: 10px;
-  flex-wrap: wrap;
-}
