--- src/Pages/Test-page/index.tsx
@@ -1,3 +1,0 @@
-import TestPage from './Test-page';
-
-export default TestPage;
