+++ src/Pages/index.ts
@@ -2,6 +2,5 @@ import GamePage from './Game-page';
 import GameResult from './Game-result';
 import LobbyPage from './Lobby-page';
 import MainPage from './Main-page';
-import TestPage from './Test-page';
 
-export { GamePage, GameResult, LobbyPage, MainPage, TestPage };
+export { GamePage, GameResult, LobbyPage, MainPage };
