+++ src/config.tsx
@@ -1,4 +1,4 @@
-export const IS_LOCAL_PROJECT = false;
+export const IS_LOCAL_PROJECT = true;
 export const IS_REMOVED_PROJECT = !IS_LOCAL_PROJECT;
 export const LOCAL_SOCKET_SERVER = 'localhost:5000';
 export const REMOVED_SOCKET_SERVER = 'https://pocker-server.herokuapp.com';
@@ -17,7 +17,7 @@ export const SETTING_SCORE_TYPE_SHORT_DEF = 'SP';
 export const SETTING_IS_SCORE_TYPE_SHORT_ERROR_DEF = false;
 export const SETTING_ROUND_TIME_DEF = 90;
 
-export const DECK_ACTION = ['?', 'Inf', 'Coffee'];
+export const DECK_ACTION = ['?', 'Inf', 'Unknown'];
 export const DECK_FIBONACCI = [
   '0',
   '1',
