--- src/data/deck.ts
@@ -1,15 +1,0 @@
-export const deck1 = ['0', '1', '2', '3', '5', '8', '13', '21', '34', '55', '89'];
-export const deck2 = [
-  '0',
-  '1',
-  '2',
-  '3',
-  '5',
-  '8',
-  '13',
-  '20',
-  '40',
-  '100',
-  '?',
-  'Coffee',
-];
